# kyber: DeriveNoise: optimize for 64 bits