benchmark                   old ns/op     new ns/op     delta
    BenchmarkDeriveNoise2-4     1343          1045          -22.19%
    BenchmarkDeriveNoise3-4     2039          1532          -24.87%